requirements.txt
fastapi
uvicorn[standard]
pymongo
SQLAlchemy
psycopg2-binary
snowflake-connector-python
databricks-sql-connector
